// app-template/openapi/types/listEventsParams.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Event Management API
 * API for Event Management System
 * OpenAPI spec version: 1.0.0
 */
import type { ListEventsStatus } from './listEventsStatus';
import type { ListEventsSortBy } from './listEventsSortBy';
import type { ListEventsSortOrder } from './listEventsSortOrder';

export type ListEventsParams = {
/**
 * Search events by name
 */
search?: string;
/**
 * Filter by event status
 */
status?: ListEventsStatus;
/**
 * Filter events starting after this date
 */
startDate?: Date;
/**
 * Filter events ending before this date
 */
endDate?: Date;
/**
 * Sort field
 */
sortBy?: ListEventsSortBy;
/**
 * Sort order
 */
sortOrder?: ListEventsSortOrder;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};